Add: New Score Calculations

+++ Assets/_Assets/u1w202012/SilCil/Scripts/NewPieceManager/IScoreCalculator.cs
@@ -4,6 +4,6 @@ namespace Unity1Week202012
 {
     public interface IScoreCalculator
     {
-        int Evaluate(IEnumerable<PieceData> pieces);
+        double Evaluate(IEnumerable<PieceData> pieces);
     }
 }
